requirements.txt
fastapi
uvicorn
pydantic
pyserial
requests
cryptography
qrcode[pil]
python-dateutil
jinja2
python-multipart
